# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ecommerce API",
        "description": "Tài liệu API cho hệ thống Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/decentralization/roles": {
            "get": {
                "tags": [
                    "Admin / Decentralization / Roles"
                ],
                "summary": "Get list of all roles",
                "description": "List all roles",
                "operationId": "listRoles",
                "responses": {
                    "200": {
                        "description": "Listed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get listed successfully!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RoleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin / Decentralization / Roles"
                ],
                "summary": "Store a new role",
                "description": "Create a new role",
                "operationId": "storeRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "role": {
                                                    "$ref": "#/components/schemas/RoleResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/decentralization/roles/{role}": {
            "put": {
                "tags": [
                    "Admin / Decentralization / Roles"
                ],
                "summary": "Update an existing role",
                "description": "Update a role",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role updated successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "role": {
                                                    "$ref": "#/components/schemas/RoleResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin / Decentralization / Roles"
                ],
                "summary": "Delete a role",
                "description": "Delete a role",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/shops/{shopId}/unlocks": {
            "post": {
                "tags": [
                    "Admin / Shops"
                ],
                "summary": "Unlock a shop that was previously locked",
                "description": "Unlock a locked shop",
                "operationId": "unlockShop",
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop unlocked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shop unlocked successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Shop already active"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleRequest": {
                "required": [
                    "name",
                    "scope"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "warehouse_manager"
                    },
                    "display_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Warehouse Manager",
                        "nullable": true
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "global",
                            "shop"
                        ],
                        "example": "shop"
                    }
                },
                "type": "object"
            },
            "RoleUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "warehouse_manager"
                    },
                    "display_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Warehouse Manager",
                        "nullable": true
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "global",
                            "shop"
                        ],
                        "example": "shop"
                    }
                },
                "type": "object"
            },
...
            "RoleResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "warehouse_manager"
                    },
                    "display_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Warehouse Manager"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "global",
                            "shop"
                        ],
                        "example": "shop"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "24-07-2025"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}